Moved DonationError, DonationForm, and DonationListForm into a new Donation directory. Removed list generation from DonationForm, refactored into new DonationList component. Created new DonationList component, that provides the list of past donations and current subscriptions.  Current subscriptions tab is currently disabled.  Styles are currently hard coded, will improve this week. ProcessingDonation was throwing warnings... it was trying to change state in the render(), so I refactored.
