+++ src/js/routes/More/ProcessingDonation.jsx
@@ -1,4 +1,4 @@
-import React, {Component} from "react";
+import React, { Component } from "react";
 import { browserHistory } from "react-router";
 import DonateStore from "../../stores/DonateStore";
 
@@ -14,7 +14,7 @@ var loadingScreenStyles = {
   alignItems: "center",
   fontSize: "30px",
   color: "#fff",
-  flexDirection: "column"
+  flexDirection: "column",
 };
 
 export default class ProcessingDonation extends Component {
@@ -23,28 +23,24 @@ export default class ProcessingDonation extends Component {
 
     this.state = {
       donation_response_received: false,
-      donation_success: false
+      donation_success: false,
     };
 
     this._onDonateStoreChange = this._onDonateStoreChange.bind(this);
   }
 
   componentDidMount () {
-//    this._onDonateStoreChange();
+    // this._onDonateStoreChange();
     this.donateStoreListener = DonateStore.addListener(this._onDonateStoreChange);
-
   }
 
-  componentWillUnmount (){
+  componentWillUnmount () {
     this.donateStoreListener.remove();
   }
 
   _onDonateStoreChange () {
-     this.setState({donation_response_received: DonateStore.donation_response_received(), donation_success: DonateStore.donation_success()});
-    }
-
-
-  render () {
+    this.setState({ donation_response_received: DonateStore.donation_response_received(),
+      donation_success: DonateStore.donation_success(), });
     if (this.state.donation_response_received) {
       if (this.state.donation_success) {
         browserHistory.push("/more/donate_thank_you");
@@ -53,6 +49,9 @@ export default class ProcessingDonation extends Component {
       }
     }
 
+  }
+
+  render () {
     return <div style={loadingScreenStyles}>
       <div>
         <h1 className="h1">Processing your Donation...</h1>
